Added Functionality to read data from file
Added a method in the Dictionary class that takes in a file path and a string of characters used as seperators of data in the file (such as commas, semi-colons, etc) as parameters. It uses a BufferedReader, FIleReader, and a StringTokenizer to read the given file. (So far I have only tested it on txt files)

If the method adds new Word Objects regardless of repeating word properties within each object. Consider handling words with multple meaning in the future. This would include not testing the if the word property is already present in the ArrayList but also the provided meaning.